Use noop control packet while waiting
The check operation essentially does an fsck of the xenstore tdb database which is expensive.  A small xenstore patch can be used to introduce a noop control function and it is preferable to use that.  No operation should be necessary here but the implementation of the wait() is wrong in this branch.  The upstream release handles this correctly but the change would be quite intrusive.